@@ -1,3 +1,3 @@
 """Python interface to the MSIS codes."""
 
-__version__ = "0.9.0"
+__version__ = "0.10.0"
